Intermediate tasks
Implemented methods to accept a local file that contains a list of family names, and performs search for each family name. The response time is recorded. At the end of each run, the average response time is calculated by dividing the total response time by the number of search requests. 

Default number of runs is 3, but can be changed by including an argument when running the program.
Cache is turned on for the 1st and 2nd runs, and turned off for the 3rd run. This is done by toggling Cache control in the client program. There should be a more elegant way of toggling the cache(perhaps in the get message), but I was unable to program it in time.